Update license and copyright boilerplates to SPDX template style
For projects originating at Red Hat, the legal team has updated the
license and copyright recommendations for source code.  The current
recommendations are extremely simple.  It is two lines:

    # Copyright AUTHORS
    # SPDX-License-Identifier: SPDX-LICENSE-ID

In the case of group projects, the recommendation is to use "The
PROJECT Project Authors" for AUTHORS.  Where PROJECT is the name of
the actual project.  In this case, "dnf".  For SPDX-LICENSE-ID, it
needs to be an approved SPDX license identifier, such as
GPL-2.0-or-later.  More information on allowed licenses available
here:

    https://docs.fedoraproject.org/en-US/legal/allowed-licenses/

One thing you cannot do is modify or strike additional copyright
statements.  So files with other Copyright statements have those
intact.  The only ones I modified were the ones of a format like
"Copyright (C) 2012-2016 Red Hat, Inc.".

Copyright statements should not include a "(C)" or "(c)" or a year or
range of years.

Signed-off-by: David Cantrell <dcantrell@redhat.com>