Use SPDX identifiers in PACKAGE-LICENSING
I also removed the note about a different license being used for
rpm/transaction.py.  After consulting with the Red Hat legal
department, the agreement was that this file should be under the GPL
version 2.0 or later.  None of the yum (or yup) code would have
existed prior to the GPL 2.0 existing and given the rest of yum being
GPL-2.0-or-later, it can be inferred that this file is also
GPL-2.0-or-later.

Signed-off-by: David Cantrell <dcantrell@redhat.com>